chore: rename stac-item-generator to stactools-item-generator
