+++ lib/index.ts
@@ -1,7 +1,7 @@
 export * from "./bastion-host";
 export * from "./database";
 export * from "./ingestor-api";
-export * from "./stac-item-generator";
+export * from "./stactools-item-generator";
 export * from "./stac-item-loader";
 export * from "./stac-api";
 export * from "./titiler-pgstac-api";
