+++ lib/stactools-item-generator/runtime/Dockerfile
@@ -10,11 +10,11 @@ ENV PATH=/tmp/.local/bin:$PATH
 
 WORKDIR ${LAMBDA_TASK_ROOT}
 
-COPY stac-item-generator/runtime/pyproject.toml pyproject.toml
-COPY stac-item-generator/runtime/src/stac_item_generator/ ${LAMBDA_TASK_ROOT}/stac_item_generator/
+COPY stactools-item-generator/runtime/pyproject.toml pyproject.toml
+COPY stactools-item-generator/runtime/src/stactools_item_generator/ ${LAMBDA_TASK_ROOT}/stactools_item_generator/
 
 RUN uv export --no-dev --no-editable -o requirements.txt && \
   uv pip install --target ${LAMBDA_TASK_ROOT} -r requirements.txt && \
   uv tool install --with requests stactools;
 
-CMD ["stac_item_generator.handler.handler"]
+CMD ["stactools_item_generator.handler.handler"]
