+++ lib/stactools-item-generator/runtime/src/stactools_item_generator/item.py
